스케줄 저장

+++ src/main/java/alz/board/service/CommentServiceImpl.java
@@ -18,12 +18,28 @@ public class CommentServiceImpl implements CommentService {
 	public CommentServiceImpl(CommentMapper commentMapper) {
 		this.commentMapper = commentMapper;
 	}
-
+	
+	// 새댓글
 	@Override
-	public CommentDTO create(CommentDTO comment) {
+	public int create(CommentDTO comment) {
 		int affectedRowCount = commentMapper.insert(comment);
-		CommentDTO createdComment = commentMapper.selectById(comment.getId()); 
-		return createdComment;
+		//CommentDTO createdComment = commentMapper.selectById(comment.getId()); 
+		return affectedRowCount;
+	}
+	
+	// 대댓글
+	@Override
+	public int create2(CommentDTO comment) {
+		comment.setCommentCnt(comment.getCommentCnt()+1L);
+		comment.setDepth(comment.getDepth()+1L);
+		
+		int updateRowCount = commentMapper.updateCnt(comment);
+		
+		if(updateRowCount!=1) return 0;
+		
+		int affectedRowCount = commentMapper.insert2(comment);
+		//CommentDTO createdComment = commentMapper.selectById(comment.getId()); 
+		return affectedRowCount;
 	}
 
 	@Override
@@ -33,23 +49,25 @@ public class CommentServiceImpl implements CommentService {
 	}
 
 	@Override
-	public List<CommentDTO> readAll() {
-		List<CommentDTO> list = commentMapper.selectAll();
+	public List<CommentDTO> readAll(Long boardId) {
+		List<CommentDTO> list = commentMapper.selectAll(boardId);
 		return list;
 	}
 
 	@Override
 	public CommentDTO updateById(Long id, CommentDTO comment) {
+		
+		int updatedRow = commentMapper.updateById(comment);
+		
 		CommentDTO searchedComment = commentMapper.selectById(id);
-		
-		searchedComment.setContent(comment.getContent());
-		
 		return searchedComment;
 	}
 
 	@Override
 	public int deleteById(Long id) {
 		CommentDTO searchedComment = commentMapper.selectById(id);
+		if(searchedComment==null) return 0;
+		
 		int affectedRowCount = commentMapper.deleteById(id);
 		return affectedRowCount;
 	}
